Revert "handle expired order and finished order"
This reverts commit 7f10cd8d05e6bd2a65a23c0c94414921ac05fc82.
