+++ proto/sao/model/query.proto
@@ -10,8 +10,6 @@ import "sao/model/metadata.proto";
 import "sao/model/model.proto";
 import "sao/model/shard_meta.proto";
 import "sao/model/expired_data.proto";
-import "sao/model/order_finish.proto";
-import "sao/model/expired_order.proto";
 // this line is used by starport scaffolding # 1
 
 option go_package = "github.com/SaoNetwork/sao/x/model/types";
@@ -52,27 +50,7 @@ service Query {
 		option (google.api.http).get = "/SaoNetwork/sao/model/expired_data";
 	}
 
-// Queries a OrderFinish by index.
-	rpc OrderFinish(QueryGetOrderFinishRequest) returns (QueryGetOrderFinishResponse) {
-		option (google.api.http).get = "/SaoNetwork/sao/model/order_finish/{timestamp}";
-	}
-
-	// Queries a list of OrderFinish items.
-	rpc OrderFinishAll(QueryAllOrderFinishRequest) returns (QueryAllOrderFinishResponse) {
-		option (google.api.http).get = "/SaoNetwork/sao/model/order_finish";
-	}
-
-	// Queries a ExpiredOrder by index.
-	rpc ExpiredOrder(QueryGetExpiredOrderRequest) returns (QueryGetExpiredOrderResponse) {
-		option (google.api.http).get = "/SaoNetwork/sao/order/expired_order/{height}";
-	}
-
-	// Queries a list of ExpiredOrder items.
-	rpc ExpiredOrderAll(QueryAllExpiredOrderRequest) returns (QueryAllExpiredOrderResponse) {
-		option (google.api.http).get = "/SaoNetwork/sao/order/expired_order";
-	}
-
-	// this line is used by starport scaffolding # 2
+// this line is used by starport scaffolding # 2
 }
 
 // QueryParamsRequest is request type for the Query/Params RPC method.
@@ -137,42 +115,6 @@ message QueryAllExpiredDataRequest {
 
 message QueryAllExpiredDataResponse {
 	repeated ExpiredData expiredData = 1 [(gogoproto.nullable) = false];
-	cosmos.base.query.v1beta1.PageResponse pagination = 2;
-}
-
-message QueryGetOrderFinishRequest {
-	  uint64 timestamp = 1;
-
-}
-
-message QueryGetOrderFinishResponse {
-	OrderFinish orderFinish = 1 [(gogoproto.nullable) = false];
-}
-
-message QueryAllOrderFinishRequest {
-	cosmos.base.query.v1beta1.PageRequest pagination = 1;
-}
-
-message QueryAllOrderFinishResponse {
-	repeated OrderFinish orderFinish = 1 [(gogoproto.nullable) = false];
-	cosmos.base.query.v1beta1.PageResponse pagination = 2;
-}
-
-message QueryGetExpiredOrderRequest {
-	uint64 height = 1;
-
-}
-
-message QueryGetExpiredOrderResponse {
-	ExpiredOrder expiredOrder = 1 [(gogoproto.nullable) = false];
-}
-
-message QueryAllExpiredOrderRequest {
-	cosmos.base.query.v1beta1.PageRequest pagination = 1;
-}
-
-message QueryAllExpiredOrderResponse {
-	repeated ExpiredOrder expiredOrder = 1 [(gogoproto.nullable) = false];
 	cosmos.base.query.v1beta1.PageResponse pagination = 2;
 }
