+++ x/model/client/cli/query.go
@@ -31,10 +31,6 @@ func GetQueryCmd(queryRoute string) *cobra.Command {
 	cmd.AddCommand(CmdShowModel())
 	cmd.AddCommand(CmdListExpiredData())
 	cmd.AddCommand(CmdShowExpiredData())
-	cmd.AddCommand(CmdListOrderFinish())
-	cmd.AddCommand(CmdShowOrderFinish())
-	cmd.AddCommand(CmdListExpiredOrder())
-	cmd.AddCommand(CmdShowExpiredOrder())
 	// this line is used by starport scaffolding # 1
 
 	return cmd
