--- x/model/client/cli/query_expired_order.go
@@ -1,77 +1,0 @@
-package cli
-
-import (
-	"context"
-
-	"github.com/SaoNetwork/sao/x/model/types"
-	"github.com/cosmos/cosmos-sdk/client"
-	"github.com/cosmos/cosmos-sdk/client/flags"
-	"github.com/spf13/cast"
-	"github.com/spf13/cobra"
-)
-
-func CmdListExpiredOrder() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "list-expired-order",
-		Short: "list all expired_order",
-		RunE: func(cmd *cobra.Command, args []string) error {
-			clientCtx := client.GetClientContextFromCmd(cmd)
-
-			pageReq, err := client.ReadPageRequest(cmd.Flags())
-			if err != nil {
-				return err
-			}
-
-			queryClient := types.NewQueryClient(clientCtx)
-
-			params := &types.QueryAllExpiredOrderRequest{
-				Pagination: pageReq,
-			}
-
-			res, err := queryClient.ExpiredOrderAll(context.Background(), params)
-			if err != nil {
-				return err
-			}
-
-			return clientCtx.PrintProto(res)
-		},
-	}
-
-	flags.AddPaginationFlagsToCmd(cmd, cmd.Use)
-	flags.AddQueryFlagsToCmd(cmd)
-
-	return cmd
-}
-
-func CmdShowExpiredOrder() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "show-expired-order [height]",
-		Short: "shows a expired_order",
-		Args:  cobra.ExactArgs(1),
-		RunE: func(cmd *cobra.Command, args []string) (err error) {
-			clientCtx := client.GetClientContextFromCmd(cmd)
-
-			queryClient := types.NewQueryClient(clientCtx)
-
-			argHeight, err := cast.ToUint64E(args[0])
-			if err != nil {
-				return err
-			}
-
-			params := &types.QueryGetExpiredOrderRequest{
-				Height: argHeight,
-			}
-
-			res, err := queryClient.ExpiredOrder(context.Background(), params)
-			if err != nil {
-				return err
-			}
-
-			return clientCtx.PrintProto(res)
-		},
-	}
-
-	flags.AddQueryFlagsToCmd(cmd)
-
-	return cmd
-}
