+++ x/model/genesis_test.go
@@ -38,22 +38,6 @@ func TestGenesis(t *testing.T) {
 				Height: 1,
 			},
 		},
-		OrderFinishList: []types.OrderFinish{
-			{
-				Height: 0,
-			},
-			{
-				Height: 1,
-			},
-		},
-		ExpiredOrderList: []types.ExpiredOrder{
-			{
-				Height: 0,
-			},
-			{
-				Height: 1,
-			},
-		},
 		// this line is used by starport scaffolding # genesis/test/state
 	}
 
@@ -68,7 +52,5 @@ func TestGenesis(t *testing.T) {
 	require.ElementsMatch(t, genesisState.MetadataList, got.MetadataList)
 	require.ElementsMatch(t, genesisState.ModelList, got.ModelList)
 	require.ElementsMatch(t, genesisState.ExpiredDataList, got.ExpiredDataList)
-	require.ElementsMatch(t, genesisState.OrderFinishList, got.OrderFinishList)
-	require.ElementsMatch(t, genesisState.ExpiredOrderList, got.ExpiredOrderList)
 	// this line is used by starport scaffolding # genesis/test/assert
 }
