+++ x/order/types/keys.go
@@ -50,8 +50,6 @@ const (
 	OrderUnexpectedEventType = "order-unexpected"
 	OrderDataReadyEventType  = "order-data-ready"
 	TerminateOrderEventType  = "terminate-order"
-	OrderExpiredEventType    = "order-expired"
-	OrderFinishEventType     = "order-finished"
 
 	NewShardEventType       = "new-shard"
 	ShardCompletedEventType = "shard-completed"
